% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{Clustering}
\alias{Clustering}
\title{Clustering}
\usage{
Clustering(Y, ClustMeth = "Hierarchical", k = 3,
  Sotadismethod = "euclidean", Pdismethod = "euclidean",
  Cdismethod = "euclidean", Ddismethod = "euclidean",
  Hdismethod = "euclidean", Hmethod = "ward.D2", Graph = T,
  VarCart = F, IndCart = F, ElbowP = F)
}
\arguments{
\item{Y}{a numeric matrix or a data frame with all numeric columns (Ex:consumers scores)}

\item{ClustMeth}{Clustering method that must be 'Hierarchical','Diana','Kmeans','Clara','Pam','Sota' or 'Som'}

\item{k}{integer, the number of clusters. It is required that 0<k<n where n is the number of observations (i.e., n = nrow(x))}

\item{Sotadismethod}{character string specifying the metric to be used for calculating dissimilarities between observations for Sota method.It could be "euclidean" or "correlation"}

\item{Pdismethod}{character string specifying the metric to be used for calculating dissimilarities between observations for PAM method.It could be "euclidean" or "manhattan"}

\item{Cdismethod}{character string specifying the metric to be used for calculating dissimilarities between observations for Clara method.It could be "euclidean","manhattan" or "jaccard"}

\item{Ddismethod}{character string specifying the metric to be used for calculating dissimilarities between observations for Diana method.It could be "euclidean" or "manhattan"}

\item{Hdismethod}{The method to calculate a dissimilarity structure as produced by dist for hierarchical method.It could be :"aitchison", "euclidean", "maximum", "manhattan", "canberra","binary" or "minkowski"}

\item{Hmethod}{the agglomeration method to be used ,should be "single", "complete", "average", "mcquitty", "ward.D", "ward.D2", "centroid" or "median"}

\item{Graph}{TRUE if you want to visualize the dendrogram (only for Hierarchical and Diana methods )}

\item{VarCart}{TRUE if you want to visualize Variables's representation}

\item{IndCart}{TRUE if you want to visualize Distribution of consumers}

\item{ElbowP}{TRUE if you want to visualize The plot of the elbow method for Hierarchical method}
}
\value{
SomCl,Graph,IndCart,VarCart,Hclustdendrogram=dend_plot,Pvar=p,Pind=p2,ElbowP=p3
}
\description{
This function serves to clustering data analysis using diverse methods and ploting diverses graphs
}
\examples{

 library(ClusteringR)
 cl=Clustering(Y=t(hedo),ClustMeth='Hierarchical',
 k=3,Hdismethod='euclidean',Hmethod="ward.D2",
 Graph=FALSE,VarCart=FALSE,IndCart=FALSE,ElbowP=FALSE )


}
