% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPM.R
\name{EPM}
\alias{EPM}
\title{External preference mapping}
\usage{
EPM(Y, X, ModelType = "Quadratic", nbpoints = 50, Graphpred = FALSE,
  Graph2D = FALSE, Graph3D = FALSE,
  statistic.Value.Type = "rsquared")
}
\arguments{
\item{Y}{:a numeric matrix or a data frame with all numeric columns (Ex:consumers scores).}

\item{X}{:a data frame with n rows (individuals) and p columns (numeric variables)}

\item{ModelType}{:Type of regression's model can be 'Vector' ,'Circular','Quadratic' or 'Eliptic'}

\item{nbpoints}{:Number of points to create the dicrete space}

\item{Graphpred}{:TRUE if you want to view the Prediction scores of one consumer heatmap,FALSE otherwise}

\item{Graph2D}{:TRUE if you want to view preferences of one consumer,FaLSE otherwise}

\item{Graph3D}{:TRUE if you want to view the external preference Map in 3D,FALSE otherwise}

\item{statistic.Value.Type}{:Extract the 'rsquared' , 'fstatistic' or 'AIC' of every regression}
}
\value{
PCA,Regression,Statistic.values
}
\description{
This function is dedicated for hedonic data segmentation
}
