\name{hedo}
\docType{data}
\alias{hedo}
\title{Hedonic data}
\description{
  This is hedonic data from a trial of 8 products within 294 customers.
}
\usage{hedo}
\format{A data frame with 8 rows and 294 variables.}
\source{Jeux de donnees en sensometrie,Jeu de donnees biscuit,<https://husson.github.io/data.html>}
\references{François Husson, M 2017, Husson posts: Jeux de donnees,Jeux de donnees en sensometrie  <https://husson.github.io/data.html>.
}
\keyword{datasets}
