
\name{senso}
\docType{data}
\alias{senso}
\title{Sensory data}
\description{
 This is sensory data from a professional trial of products
}
\usage{senso}
\format{A data frame with 192 rows and 294 variables:}
\source{Jeux de donnees en sensometrie,Jeu de donnees biscuit,<https://husson.github.io/data.html>}
\references{François Husson, M 2017, Husson posts: Jeux de donnees,Jeux de donnees en sensometrie  <https://husson.github.io/data.html>.
}
\keyword{datasets}

