% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedo.R
\docType{data}
\name{hedo}
\alias{hedo}
\title{Hedonic data}
\format{A data frame with 8 rows and 294 variables.}
\usage{
hedo
}
\description{
This is hedonic data from a trial of 8 products within 294 customers.
}
\examples{
\dontrun{
 library(ClusteringR)
 cl=Clustering(Y=t(hedo),ClustMeth='Hierarchical',
 k=3,Hdismethod='euclidean',Hmethod="ward.D2",
 Graph=T,VarCart=F,IndCart=F,ElbowP=F )
}

}
\references{
François Husson, Jeux de donnees,
Jeux de donnees en sensometrie, <https://husson.github.io/data.html>.
}
\keyword{datasets}
