% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senso.R
\docType{data}
\name{senso}
\alias{senso}
\title{Sensory data}
\format{A data frame with 192 rows and 294 variables:}
\usage{
senso
}
\description{
This is sensory data from a professional trial of products
}
\examples{
\dontrun{
library(ClusteringR)
EPM(Y=t(hedo),X=senso,ModelType='Quadratic',
nbpoints=50,Graphpred=FALSE,Graph2D=FALSE,
Graph3D=FALSE,statistic.Value.Type='rsquared')}
}
\keyword{datasets}
